requirements.txt
ffmpeg-python
python-telegram-bot
aiohttp
aiogram
heroku3
motor
cryptg
tgcrypto
ethon
Pillow>=5.3.0
aiofiles~=0.8.0
pymongo[srv]
pyrogram
requests
telethon
dnspython
python-decouple